pyproject.toml
[project]
name = "comicif-backend"
version = "0.1.0"
description = "Django backend for comic-style photo processing with background replacement"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.4.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "gunicorn>=23.0.0",
]
